Refactored Scheduling and Notification system
