+++ src/main/java/lux/fm/bookingservice/service/NotificationService.java
@@ -6,4 +6,6 @@ import org.springframework.stereotype.Service;
 public interface NotificationService {
     void notifyUser(Long id, String message);
 
+    void notifyAllUsers(String message);
+
 }
