+++ src/main/java/lux/fm/bookingservice/service/SchedulingService.java
@@ -1,0 +1,5 @@
+package lux.fm.bookingservice.service;
+
+public interface SchedulingService {
+    void checkExpiredBookings();
+}
